projects/finance_nlp_pipeline/requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
transformers>=4.40
torch>=2.1
openai>=1.0
datasets>=2.18
mlflow>=2.10
fastapi>=0.110
uvicorn>=0.27
netcal>=1.3
rouge-score>=0.1.2
bert-score>=0.3.13
shap>=0.44
python-dotenv>=1.0
jupyter>=1.0
joblib>=1.3
